+++ src/main/java/seedu/savenus/model/ReadOnlyMenu.java
@@ -1,6 +1,6 @@
 package seedu.savenus.model;
 
-import javafx.beans.property.FloatProperty;
+import javafx.beans.property.DoubleProperty;
 import javafx.beans.property.IntegerProperty;
 import javafx.collections.ObservableList;
 import seedu.savenus.model.food.Food;
@@ -23,9 +23,9 @@ public interface ReadOnlyMenu {
     Wallet getWallet();
 
     /**
-     * Returns the user's {@code RemainingBudget}'s {@code FloatProperty}.
+     * Returns the user's {@code RemainingBudget}'s {@code DoubleProperty}.
      */
-    FloatProperty getRemainingBudgetProperty();
+    DoubleProperty getRemainingBudgetProperty();
 
     /**
      * Returns the user's {@code DaysToExpire}'s {@code IntegerProperty}.
